Changed sound implementation

+++ mcnative-runtime-api/src/main/java/org/mcnative/runtime/api/protocol/packet/type/sound/MinecraftSoundEffectPacket.java
@@ -66,7 +66,7 @@ public class MinecraftSoundEffectPacket implements MinecraftPacket {
         return pitch;
     }
 
-    public void setPitch(byte pitch) {
+    public void setPitch(float pitch) {
         this.pitch = pitch;
     }
